requirements.txt
pandas==1.4.1
scikit-learn<=1.0.2
joblib==1.2.0
fastapi[all]==0.86.0
uvicorn[standard]==0.18.3
sqlmodel==0.0.8
pymysql==1.0.2
python-dotenv==0.21.0
psycopg2-binary==2.9.5
passlib==1.7.4
bcrypt==4.0.1
PyJWT
